First dev release of vscode-ws-jsonrpc from this repository
- Aligened code with node 16 api
- Moved CHANGELOGs to packages and the top-level one links to them
- Aligned tsconfigs

+++ packages/vscode-ws-jsonrpc/src/server/launch.ts
@@ -10,14 +10,16 @@ import { StreamMessageReader, StreamMessageWriter, SocketMessageReader, SocketMe
 import { IConnection, createConnection } from "./connection";
 import { IWebSocket, WebSocketMessageReader, WebSocketMessageWriter, IWebSocketConnection } from '../socket';
 
-export function createServerProcess(serverName: string, command: string, args?: string[], options?: cp.SpawnOptions): IConnection {
-    const serverProcess = cp.spawn(command, args, options);
+export function createServerProcess(serverName: string, command: string, args?: string[], options?: cp.SpawnOptions): IConnection | undefined {
+    const serverProcess = cp.spawn(command, args || [], options || {});
     serverProcess.on('error', error =>
         console.error(`Launching ${serverName} Server failed: ${error}`)
     );
-    serverProcess.stderr.on('data', data =>
-        console.error(`${serverName} Server: ${data}`)
-    );
+    if (serverProcess.stderr !== null) {
+        serverProcess.stderr.on('data', data =>
+            console.error(`${serverName} Server: ${data}`)
+        );
+    }
     return createProcessStreamConnection(serverProcess);
 }
 
@@ -37,8 +39,12 @@ export function createSocketConnection(outSocket: net.Socket, inSocket: net.Sock
     return createConnection(reader, writer, onDispose);
 }
 
-export function createProcessStreamConnection(process: cp.ChildProcess): IConnection {
-    return createStreamConnection(process.stdout, process.stdin, () => process.kill());
+export function createProcessStreamConnection(process: cp.ChildProcess): IConnection | undefined {
+    if (process.stdout !== null && process.stdin !== null) {
+        return createStreamConnection(process.stdout, process.stdin, () => process.kill());
+    } else {
+        return undefined;
+    }
 }
 
 export function createStreamConnection(outStream: stream.Readable, inStream: stream.Writable, onDispose: () => void): IConnection {
